moved BatchSettings into a separate file
